+++ cpp/src/strings/split/split.cu
@@ -25,6 +25,7 @@
 #include <cudf/strings/string_view.cuh>
 #include <cudf/strings/strings_column_view.hpp>
 #include <cudf/utilities/error.hpp>
+
 #include <strings/split/split_utils.cuh>
 
 #include <thrust/binary_search.h>  // upper_bound()